Move to generic HTTP interface in management.
Guardian Frontend 24998149: Meet web app specifications
Want to use management code in the Frontend Play applications and the
dependency on Servlet API needs to be unpicked for this. The management
library is relatively Servlet API independant on the request side but
it is a bit tight on the response side.

Much of this is about retaining the Lift-style response wrappers around
an intermediate HTTP abstraction to replace the Servlet API
HttpServletResponse.

Review fixups.

* Fix for encode typo.
* Elabourated documentation.
* Rename HttpServletHttpR* -> ServletHttpR*