Enables pan-app request metrics in the example-play app.
I think this is the canonical way to enable these metrics. It's worth
noting that the metrics previously set up in Global.scala weren't actually
getting displayed anywhere, as they weren't being passed into the
constructor of the management StatusPage.

Also use internal Management server (port 18080) for example-play app -
this makes it easy to do request metrics that cover the entire app,
but not count requests to the management urls themselves.