Use devproxy.gul3.gnl instead of decommissioned devscreen.gudev.gnl
My hope is this will let the build resolve this dependency:

org.avaje.ebeanorm#avaje-ebeanorm-agent;[3.1,4)

...which is currently killing my build.

---------- Forwarded message ----------
From: Michael Brunton-Spall <michael.brunton-spall@guardian.co.uk>
Date: 12 December 2012 12:48
Subject: devscreen.gudev.gnl is decommissioned
To: "dig.dev@guardian.co.uk" <Dig.Dev@guardian.co.uk>


So it appears that devscreen.gudev.gnl stopped working sometime in the last few hours.

This box is incrediably unhealthy and we've been looking for a way to decommission it anyway, so I am not going to fix it.

For various reasons this box ran a squid caching proxy, and it might be possible that your machine is configured to use it as a proxy for SBT or Maven and so forth.  You'll need to stop that and find out why you were using a proxy and work out if there is a better solution for you.

It might also be the case that your teramcity build sets the http proxy to be devscreen.gudev.gnl, needless to say it should not do this.  For TeamCity's use (and teamcity's use only), there is a proxy devproxy.gul3.gnl on port 3128

Michael Brunton-Spall
Developer Advocate
guardian.co.uk